requirements.txt
# Core agent dependencies
redis==5.0.1
httpx==0.27.2
uvloop==0.19.0; platform_system != "Windows" and python_version < "3.13"
loguru==0.7.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
aiohttp==3.9.1

# Manager module dependencies
PyYAML==6.0.1
jsonschema==4.22.0
